microthreads for mono-4.2.1.102
The MMRUThread class provides co-routines or stack switching for efficient
user-mode thread switching in a co-operative environment.  They are an
improvement over the supplied Continuations class in that they do not do a
memcpy() and memset() to switch stacks, just a stack pointer register swap.
This also allows the microthreads to be run from different macrothreads such as
from a worker thread pool, as there is no 'mark top of stack' call that binds a
microthread to a particular macrothread.

See UTHREADAPI.txt for details.

Currently only implemented for x86 32-bit and 64-bit.